Refactor: 인터셉터 파일 위치 수정 - BE [#47]
