+++ BE/src/main/java/todo7/BE/config/CommonWebConfig.java
@@ -2,10 +2,9 @@ package todo7.BE.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import todo7.BE.utils.AuthCheckInterceptor;
+import todo7.BE.web.AuthCheckInterceptor;
 
 @Configuration
 public class CommonWebConfig implements WebMvcConfigurer {
